// app/users/api/user.ts
export class User1 {
    id: number;
    username: string;
    password: string;
    first_name: string;
    last_name: string;
    email: string;
    bio: string;
    high_school: string;
    courses: string [];
    grad_year: number;
    imagePath: string;
}

export class User2 {
    id:number;
    first_name: string;
    last_name: string;
    bio: string;
    classes:string [];
    photo: string;
    high_school:string;
}